Introduce batch updater for sandbox code samples
It allows to automatically update the code samples of the
configured set of sandbox repositories and open pull requests
against each of them.

As before, the base code samples live in this very template
repositories while the examples are defined in metadata defined
in the README files of each sandbox repository.

Signed-off-by: Florent Biville <florent.biville@neo4j.com>